fix(build): 适配 NDK 28 编译环境
- fftools_cmdutils.c: 修复 INT64_MIN/INT64_MAX 隐式转换警告，添加显式 (double) 强转
- mobileffmpeg.c: 将 VLA 常量改为 #define 宏定义，修复 NDK 28 编译错误
- Android.mk: 移除 c++_shared 模块依赖声明（NDK 28 自动处理），移除 -Werror 避免警告阻断编译
- android-common.sh: target triple 追加 API level 后缀，移除 -fno-integrated-as，gold 链接器改为 lld

+++ android/app/src/main/cpp/fftools_cmdutils.c
@@ -358,7 +358,7 @@ static int write_option(void *optctx, const OptionDef *po, const char *opt,
     } else if (po->flags & OPT_BOOL || po->flags & OPT_INT) {
         *(int *)dst = parse_number_or_die(opt, arg, OPT_INT64, INT_MIN, INT_MAX);
     } else if (po->flags & OPT_INT64) {
-        *(int64_t *)dst = parse_number_or_die(opt, arg, OPT_INT64, INT64_MIN, INT64_MAX);
+        *(int64_t *)dst = parse_number_or_die(opt, arg, OPT_INT64, (double)INT64_MIN, (double)INT64_MAX);
     } else if (po->flags & OPT_TIME) {
         *(int64_t *)dst = parse_time_or_die(opt, arg, 1);
     } else if (po->flags & OPT_FLOAT) {
